experiments/requirements.txt
numpy
Pillow
scipy
dlib==19.22
opencv-python
torchvision
matplotlib
pandas